requirements.txt
alembic==1.12.0
anyio==4.0.0
bcrypt==4.0.1
black==23.3.0
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
distlib==0.3.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.95.2
FastAPI-SQLAlchemy==0.2.1
filelock==3.12.4
greenlet==3.0.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.11.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.13
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
ruff==0.0.272
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.22.0
virtualenv==20.24.5
